. Rsid tags were not removed
The following rsid tags did not belong to the list of rsid tags to be removed : styrsid and rsidroot

The space after an rsid tag is usually a regular space which belongs to the plain text of the document. It was undully removed, thus causing non-adjacent characters to become adjacent.